+++ Changelog.md
@@ -1,3 +1,7 @@
+### aws_csshx 0.1.4 2012-05-21
+
+ * Better error handling (Eric Lubow)
+
 ### aws_csshx 0.1.3 2012-05-05
 
  * Ability to add individual servers on the command line (Eric Lubow)